cli/pyproject.toml
[project]
name = "wb"
version = "0.1.0"
description = "Workbench CLI for knowledge management"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
wb = "wb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
